+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added constants to `PdfPngParser` class.
 - Updated PhpCsFixer configurations for line span in PHPDoc blocks.
 
 ## 4.0.1 - 2026-02-10